chore(auth-tests): create application factory util, rename files
